requirements.txt
google-api-python-client
python-dotenv
pandas
numpy
vaderSentiment
snscrape
tqdm

